Fixed data race in GOSoundOrganEngine::GetMeterInfo (#2486)
GetMeterInfo() was called from the GUI thread while accessing
m_AudioOutputTasks, which could be concurrently resized or cleared
by the control thread in SetAudioOutput() or ClearSetup().

Meter data is now collected from tasks in NextPeriod() (audio thread)
and stored atomically into m_MeterInfo (vector of std::atomic<float>).
GetMeterInfo() only reads m_MeterInfo under m_MeterMutex, which also
protects against concurrent resize in SetAudioOutput().



Added comments to SetAudioOutput explaining why m_MeterInfo is replaced



Added assert in NextPeriod to detect m_MeterInfo out-of-bounds writes



Used exchange(0) to atomically read and reset m_UsedPolyphony



Fixed meter data race and peak tracking regression (review #2485)

- Add accumulate_max_to_atomic() helper (fetch_max is C++26 only)
- NextPeriod: use atomic max instead of store so peaks accumulate
  between GUI polls; skip task[0] (downmix/recorder, not a real output)
  to prevent out-of-bounds write when audio_outputs is empty
- GetMeterInfo: reset m_MeterInfo via exchange(0) so next poll starts
  fresh; document no-race with NextPeriod (SetAudioOutput runs before
  the audio callback starts)
- Update m_MeterInfo comment to describe the new accumulation model



Used atomic pointer to meter vector for cache visibility after Stop/Start

Replace std::vector<std::atomic<float>> m_MeterInfo with:
- mp_MeterInfo (unique_ptr): owns the vector
- p_MeterInfo (atomic<vector*>): always == mp_MeterInfo.get()

SetAudioOutput() allocates a new vector under m_MeterMutex and publishes
it via p_MeterInfo.store(release). NextPeriod() loads p_MeterInfo with
acquire, guaranteeing that after a Stop/Start cycle the audio callback
thread sees the new vector's size and data — not a CPU-cached old pointer.
GetMeterInfo() reads *p_MeterInfo under m_MeterMutex as before.



Documented why destroying old meter vector is safe in SetAudioOutput

Audio callbacks are paused between StopSoundSystem() and StartSoundSystem(),
so no audio thread holds a pointer to the old vector when it is destroyed.



Addressed review #2486: polyphony back to m_UsedPolyphony, store before destroy

- Restore m_UsedPolyphony (atomic_uint) for polyphony tracking; mp_MeterInfo
  now holds only per-channel peaks (size = real output channels, no slot [0])
- Use accumulate_max_to_atomic (now templated) for m_UsedPolyphony updates
- SetAudioOutput: reallocate mp_MeterInfo only when channel count changes;
  store new pointer (release) before destroying old vector
- GetMeterInfo: assert(hardPolyphony > 0); prepend polyphony ratio from
  m_UsedPolyphony.exchange(0); use pointer increment + range-for
- Add shared_ptr note in header explaining why it is unsuitable



Renamed accumulate_max_to_atomic to atomic_fetch_max_relaxed, use relaxed ordering



Addressed review #2486: clarify meter loop rationale, fix stale comment

- Expand comment on the i=1 loop to explain that task[0] was never
  included in the meter display, and that GOFrame::OnMeters guards
  against size mismatches so a transient channel-count change safely
  skips one GUI tick instead of crashing.
- Fix stale function name in the .h ownership comment:
  accumulate_max_to_atomic → atomic_fetch_max_relaxed.



Addressed review #2486: explicitly zero-initialize meter atomics

Prior to C++20, std::atomic default construction does not zero-initialize
the value. Add an explicit store(0.0f) loop after allocation.

+++ src/grandorgue/gui/frames/GOFrame.cpp
@@ -764,7 +764,8 @@ void GOFrame::OnSize(wxSizeEvent &event) {
 
 void GOFrame::OnMeters(wxCommandEvent &event) {
   if (m_isMeterReady) {
-    const std::vector<double> vals = r_SoundSystem.GetEngine().GetMeterInfo();
+    const std::vector<float> vals = r_SoundSystem.GetEngine().GetMeterInfo();
+
     if (vals.size() == m_VolumeGauge.size() + 1) {
       m_SamplerUsage->SetValue(33 * vals[0]);
       for (unsigned i = 1; i < vals.size(); i++)
